Removed unused birthdate fields from form.
