+++ src/auth/components/login/Login.tsx
@@ -6,31 +6,12 @@ import authenticate from '../../authenticate';
 import PageLayout from '../../../common/layout/PageLayout';
 import styles from './Login.module.css';
 import BirthdateForm from '../birthdateForm/BirthdateForm';
-import Button from '../../../common/button/Button';
 
 type Props = {};
 
 function Login(props: Props) {
   const [showManualRegistration, setShowManualRegistration] = useState(false);
   const { t } = useTranslation();
-
-  // TODO move to own file
-  const getCookie = (cookieName: string) => {
-    const name = cookieName + '=';
-    const decodeCookie = decodeURIComponent(document.cookie);
-    const cookieArray = decodeCookie.split(';');
-
-    for (let i = 0; i < cookieArray.length; i++) {
-      let cookie = cookieArray[i];
-      while (cookie.charAt(0) === ' ') {
-        cookie = cookie.substring(1);
-      }
-      if (cookie.indexOf(name) === 0) {
-        return cookie.substring(name.length, cookie.length);
-      }
-    }
-    return '';
-  };
 
   const checkBirthDate = (birthDate: string) => {
     const age = differenceInYears(new Date(), new Date(birthDate));
@@ -39,7 +20,7 @@ function Login(props: Props) {
       setShowManualRegistration(true);
     } else {
       document.cookie = `birthDate=${birthDate}`;
-      //authenticate();
+      authenticate();
     }
   };
